Move ParseRequest to an apache2 package.
