+++ pkg/loghttp/push/push_test.go
@@ -1,4 +1,4 @@
-package distributor
+package push
 
 import (
 	"bytes"
